requirements.txt
# Backend Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
Pillow==10.1.0
aiofiles==23.2.1
httpx==0.25.1
boto3==1.34.0